Skip hidden folders
Move utility methods
