+++ lib/index.js
@@ -1,30 +1,7 @@
 const fs = require("fs");
 const path = require("path");
 
-function getDefaultFileTitle(filePath) {
-  const fileSegments = filePath.split("/");
-  const fileName = fileSegments[fileSegments.length - 1];
-  const ext = path.extname(fileName);
-  return fileName.split(ext)[0];
-}
-
-function getMarkdownTitle(filePath) {
-  const content = fs.readFileSync(filePath, "utf8");
-
-  const titleMatch = content.match(/^#\s+(.*)/m);
-
-  if (titleMatch) {
-    return titleMatch[1].trim();
-  }
-  return getDefaultFileTitle(filePath);
-}
-
-function getFileTitle(filePath) {
-  if ([".MD", ".md"].includes(path.extname(filePath))) {
-    return getMarkdownTitle(filePath);
-  }
-  return getDefaultFileTitle(filePath);
-}
+const { shouldSkipFolder, getFileTitle } = require("./utils");
 
 function getTableOfContents({ dirPath = process.cwd(), extensions = [] }) {
   try {
@@ -45,6 +22,9 @@ function getTableOfContents({ dirPath = process.cwd(), extensions = [] }) {
         const stat = fs.statSync(fullPath);
 
         if (stat.isDirectory()) {
+          if (shouldSkipFolder(file)) {
+            continue;
+          }
           toc += `${" ".repeat(level * 2)}* **${file}**\n`;
           readDirectory({
             directory: fullPath,
